reafctor(module): Renames module to harvester
Since module is such a generic name for the folder.
Also renames entrypoint fetch_osm_pumps to main.
Moves fetch(osm_pumps) function to its own file.
